Added the facility to filter and order various lists.
Each controller currently has a mapping of "normal" to node-path names
from which it generates parameters for the REST API. This stuff should
ultimately be moved closer to the data models, since it's very fragile
and sensitive to changes in the model. It's also a bit of a hack anyway
but it'll do until we get Solr set up and working.

Various list template renderings have been cleaned up as well.