Big refactoring to passing of user parameter from Actions and views.
Previously, some views accepted an Option[UserProfile] and others
(mostly those where someone had to be logged in) a UserProfile. Now,
an Option[UserProfile] is passed everywhere, no matter if it's always
full or not. This removes some ugliness where the we had to convert to
an implicit option get to use the AsyncRest error wrapper, and makes
re-using templates a lot more versatile, since we can just use the
always-in-scope implicit userOpt everywhere.