Added CSRF protection to all POST forms.
As expected, most of the tests fail but, weirdly, not all of them.
The add/remove user from group tests, for example, pass, when they
shouldn't. If we disable CSRF in tests by passing a non-filtered
Global object to the FakeApplication constructor, then there are
crashes only when a POST form fails validation and is re-rendered.