Font changes.
The 'Lato' google web font looked bad with a lot of gylphs we need,
such as Latvian diacritics. No other (free) web fonts seem to be
available that look nice for headings and have wide glyph support.
For the time being the heading font as been changed to bootstrap
default serif (with the body font still sans-serif.)