Beginnings of better action builders, based on Play2-auth 0.13:
 - use `ActionTransformer`, `ActionRefiner`, and `ActionFilter` to
   break up the multifarious functions of `userProfileAction`
 - deprecate some of the existing partial functions

NB: We can't yet use wrapped requests to replace the generic item-based
action  functions here because the Play action builders don't seem to
work well when the wrapped request has more than one generic type (in this
case the type of its item.)

We _can_ do this, however, in the generic controller traits because the
type is on the class level.

TODO: Fix numerous deprecation warnings and flesh out more of the item-
based generic controller traits.