Abandoning Yahoo OAuth2 unfinished, but committing useful refactoring to OAuth2 flow.
Yahoo seems to be problematic at the moment because whilst two-thirds of the OAuth2
flow is working, fetching user info isn't due to a) permissions, and b) YQL difficulties.

However, this includes useful refactoring to the OAuth2LoginHandler and associated classes
to make the overall flow more clear and providers more generic.