Fix a major issue with the refactored search query builder that broke pagination.
It seems that Scalikesolr adds start and rows parameters in the group params
in addition to the default. This just happened to work before, presumably due
to the order of the params in the query, but now we convert them to a map
it stopped working. For the time being have explicitely specified start and
row params to the group settings, which at least means the duplicate params
are the same.

Another vote for phasing our Scalikesolr for our usage.