Add handling for 404s to the action helper we typically use to fetch stuff.
These are now logged as warnings with the referrer, instead of propagating to
the global error handler and being logged as an error.

There are still other code paths where the backend can trigger 404s, but these
are typically more pernicious and likely indicative of problems with the data,
so it's best to retain the current error handling here.