Set logback conf file path explicitly in build for testing.
This shuts up most instances of the multiple instances on classpath
problem that results from both Play and Neo4j packaging a logback.xml,
albeit not all of them. SBT seems to prompt some logback output
of its own somewhere outside of our projects and I'm not entirely sure
where this is coming from. Still, it's an improvement.