Change markdown rendering from a static threadlocal to injected component.
There is some overhead associated with instantiating a Pegdown processor,
so to avoid this we've previously cached a threadlocal instance. Now,
however, it's nicer to make it an injectable component. Unfortunately,
this requires changes approximately 103838848 individual view templates
to add the implicit renderer var, so some consolidation would be nice.