First pass at fixing the messages problems caused by renaming types.
For the moment, the bare type prefix is the same (old lowercase values)
but hardcoded where it is used. Where the type is not a prefix (e.g. more
of an enumeration) it's uppercased, matching the type. The reason being that
the former not commonly used in a programmatical way, whereas the latter are.

Ultimately a better solution is required, preferably one which makes the
distinction between the type and its description.

Probably many messages bugs still remaining as a result of this.