Fixes and refactoring of backend method forwarding, etc.
Previously there was redundancy in the way we fetched generic data
from the backend (via GET methods). This has now been consolidated
into one controller (renamed `Data`). It is now also possible to
forward a backend method that contains a '/', which was broken before.
This allows exposing the new (WIP) schema export.

Additional fixes to some virtual unit admin templates, which none-
theless continue to be a massive disaster.