Rework asset pipeline to ensure JS is compressed.
Also tweak some API doc templates and fix an issue where the
search indexer produced too much progress output for the browser
to handle.

CSS is still not compressed automatically because sbt-css-compress
slows down development so much when it runs on every reload - need
to find a way to resolve this.

Fixes #974.