Reimplement access point management widget in Vue.js
This is pretty much the exact same functionality as before but
done in a hopefully more maintainable manner. In the future we
should refactor how the whole feature is implemented across the
whole stack, with appropriate changes on the back end.

Since behaviour is subtle the old widget has been kept around until
we can verify that everything works as expected in production.