Change behaviour of vocabs and sets pages to query contents.
Previously the vocabularies and authoritative sets pages allowed
users to search only (promoted) vocabs and sets. However this is pretty
useless and counter-intuitive since there are only a handful of either.

This change returns results for their *contents* instead of the
holders.

The main problem here is that it's different behaviour to the
other search pages.