Prototype a rough way to upload file data for institutions.
This could, one day, perhaps be ingest data. At the moment it is
mainly just a test of using multiple S3-compatible Cloud APIs.
Ideally uploaded data would go directly to the Cloud rather than
via our server but more work is needed for facility that, if it's
even possible with DO. Note that DO does not seem to support versioning
yet, which is a bit of a showstopper.