Save an import log every time an import is committed
This is opposed to only saving one when an event is saved, i.e.
work has been done. This allows more easily figuring out the state
of files and lays the groundwork for better retrying of errors etc.

Database schema changed to a native fkey on the import_log table.