Add a way of 'harvesting' a set of files via a list of URLs
Since there was too much duplication in the different methods for
CRUDing harvest configurations these have been combined into a
single controller which dispatches to the appropriate service as
necessary.

On the frontend there is also a new way of resizing certain modal
dialogs and some other UI tweaks.

Partial fix for #1403