Fix XSS vulnerability via unsafe rendering of objects in link text
We (sloppily and unnecessarily) wrapped object toStrings in Html() when
rendering links. This resulted in a vulnerability where a user could put
XSS in their name and it would be rendered via name links.

This is fixed via a better separation between text and html content
passed to the link renderering partial and by not arbitrarily assuming
names are safe. Additionally, the highlighter has another method for
explicitely non-tagged text which is sanitised.

TODO: more auditing of the existing highlighting method, which needs
to handle mixed content. Possibly server-side scrubbing of some fields.

Fixes #1482