Allow importing EADs nested under specified fonds on a dataset
Typically the fonds ID specifier is for `sync` update, where the
info for the fonds is in the EAD to be imported. But sometimes we
want to import items _under_ an existing fonds. This new `nest`
option on the dataset allows this.

Also add some tests for the ImportConfigs controller and better
validation for the fonds ID.